dataretention: fix build on platforms where Statfs_t.Bavail is int64 (#896)
On FreeBSD (and some other targets) syscall.Statfs_t.Bavail is int64
while Blocks/Bfree are uint64, so the available = st.Bavail * bsize
multiplication failed to compile. Convert to uint64 explicitly, matching
the existing bsize conversion.

Follow-up to 835e09cb8.

+++ pkg/history/dataretention/statfs.go
@@ -23,7 +23,7 @@ func diskCapacity(dataDir string, dbUsedBytes uint64) (capacity uint64, otherUse
 	if diskUsed > dbUsedBytes {
 		otherUsed = diskUsed - dbUsedBytes
 	}
-	available = st.Bavail * bsize
+	available = uint64(st.Bavail) * bsize
 	ok = true
 	return
 }
